Adjust plugin configuration files to load.
According to the plugins manager dein, the normalized names are used to
load the configuration.
Therefore a bunch of configuration files for the plugins have been
renamed to get loaded by this name.
Simplify the plugin manager configuration by removing the name
definitions for the plugins as before.
Adjust some plugin loads to be more lazy. Add some TODOs, where it does
not work so far (merged plugin folder seems to be the reason)