Add feature to fix window layout.
Add new script 'layout' which gets automatically loaded.
A dictionary defines the fixed windows, which should be correct by
a variant of properties.
When calling the function, each entry will be searched for and if such
window could been found, the properties are adjusted.
Add a mapping for windows to call this function.
Define the window height variable for the YankRing plugin.

Besides make all Git mappings silent.