move all plugin registration to optional packages
The registration of a plugin must not be done on every startup of NeoVim. Only
when the plugin manager needs to act on the plugins. Therefore a lot of `start`
packages could be completely removed or at least one plugin file. Only lazy load
plugins with mappings must have a `start` package now. But every single plugin
has a `opt` package now. The Lua file with the plugin registration could gets
placed into the root of each `opt` package, making NeoVim ignore this file. It
will be manually sourced via a `runtime` pattern call when needed.