Hop: switch mapping keys
On my custom keyboard, the space key has a couple of special functionalities. In
result double tapping it doesn't work properly to trigger NeoVim mappings.
The backspace key is also very comfortable under the thumb and is more easy. In
normal and visual mode the key just works like `h` or `<left>` anyways. After it
saves even a keystore, making it easier to trigger the mapping.