rename all start packages without underscore
There was an issue in the past that when loaded an optional package for
a name that equals a start package, the start package gets also loaded
again. I solved this by adding a trailing underscore to all start
packages.
There was an open issue for this at NeoVim. Apparently this has
been resolved according to recent tests.