ufo: add as new plugin
Primarily for the syntax highlighting of the fold-line.
The LSP provider is a bit slow on start which is annoying. I also have
concerns if the language server should be responsible or if this is
rather a treesitter topic. I'll experiment.

This also removed a lot of code of myself and allows to simply the
setup.