codelens: fix auto command to refresh lenses
The `pcall` was somehow not working. In general there seems to be an
issue if a lens refresh gets called and fails. Looking into the NeoVim
codebase, there is a mechanism to prevent concurrent polls. It seems
like this mechanism is broken, though it actually should be able to
handle errors safely.

As now the code lenses actually work, the nightlights were adapted. The
former colors where meant to be screamy, so I see when the first servers
support lenses.