improve startup time via plugin manager
It is simply not necessary to run the plugin manager setup on every
single startup. Rather only when the plugins are updated and installed.
The only missing problem is the custom package path and how it gets
handed down. To avoid this issue, I'm using a default package path now.
Improves startup time by ~15ms (~6%).