add utilitiy to receive icons from diagnostics signs
This functionality is already very long existing. But I haven't pushed it yet
because it was related to a local experimental feature. Now I realized that it
was already used all the time by the bufferline plugin. Recognized this on
a fresh setup on a new PC where the access to this function fails.