Heirline: fix loading issue with colors
When Packer was re-compiling the loading script, the order of the
Heirline plugin and the one to set the colorscheme changed randomly. The
problem is that Heirline tries to access some custom highlight groups.
If they don't exist, the whole NeoVim startup kinda "fails".
The "workaround" was to just re-compile it again and again until the
order was correct again.

To solve this more properly now, Heirline gets loaded lazily now. The
issues was just that the `ColorScheme` event wasn't working as it is
somehow too early in the event chain and the Packer script isn't sourced
yet.
For now, the `VimEnter` event seems to work reliably because here all
`plugin/*` scripts are already sourced.