nvim-cmp: prevent unintended snippet expansion
I started to use LuaSnip its multi-snippet feature where I especially
use the same snippet with automatic and manual trigger.
The default behavior of nvim-cmp on selecting a completion menu entry is
to insert its value into the buffer. An additional completion
confirmation triggers additional features like imports, snippets, etc.
Because the completion entry for the manual snippet is the same as the
automatic, as soon as the first snippet completion item is selected, it
automatically triggers.
To prevent this, that selection behavior is alternated so snippet items
always require confirmation.