feat(plugin/nvim-cmp): add automatic confirmation
When selecting a completion entry from the menu and just continue typing
after the value was inserted, the additional text edits of this
completion entry are not applied. That requires a manual confirmation of
the selected entry. This can be quite annoying.

This feature remembers if a completion entry was manually selected. When
the completion menu closes, it will trigger a confirmation of that
entry. This will trigger effects like automatic imports etc.
In case the window closes without a selection, nothing happens. In case
of a manual confirmation, the selection will be cleared first, to avoid
double confirmation.