Fix(inlay_hint): solve toggling issues
After turning off inlay hints with the custom toggle command, hints
start to appear again after some time. That was due to the
auto-commands, which still trigger them again. As they can't determine
the toggling state, a proper solution is to remove the auto-commands
when turning the hints off. That required some restructuring of code.