Feat(lsp/middleware): create only configured handlers
This avoids to create a handler for every single LSP method due to the
meta-table.

While this change in itself just seems right, it has been introduced to
fix a certain bug. Somehow NeoVim only fully parses a message body by a
server if there is a handler for the method. As the old implementations
returns a handler for any possible method, this was always the case. But
there are message by servers that NeoVim core itself does not support
and fails during parsing the body. That was especially the case for the
`workspace/diagnostic/refresh` message. Unfortunately some server seem
to ignore the `workspace.diagnostics.refreshSupport` capability and send
these messages anyway.