Feat(lsp/server): improve starting of servers
Add wrapper for `vim.lsp.start` which evaluates the configured root
directory and the server command to determine if a server should be
started.

Remove the need to pass the file-type parameter when building the start
option. It is a repeating pattern that the current buffer file-type is
used for transitive effects. Therefore this is the default now
automatically.