Chore: remove diagnostic sign integrations
Unfortunately NeoVim core removed regular signs for diagnostics.
Thereby it got hard to access them. The only way seem to read the
diagnostic configuration. But that only works if it was configured by
the user. So there is no access to the defaults possible.
In the end I decided to drop the integrations as they don't matter much
to me anymore.