feat(telescope.nvim): align with native mappings
This isn't really clean. Usually the lower case character is for more
local/smaller scopes. The same character in upper case for a related
broader scope. Unfortunately does the native mapping use the upper case
character for a local scope. There is no mapping for the other scope
natively yet.