feat: introduce and apply custom lazy loading by commands
The goal is to move away from any plugin manager, but make use of the
new default in-built one. However, lazy loading capabilities are
missing. In most cases, this is quite simple logic. In that context,
some simple lazy loading by a Vim Ex command has been added an applied.

This removes the number of features used by the current plugin manager.
The next steps are to remove one feature after another, until the
in-built plugin/package manager can be used.

As a side-effect of this, plenty of key mappings have been transformed
into Lua. This was a long term goal anyway. It fit with the current
uncertain placement of where to establish the placeholder commands for
lazy loading.
Furthermore, all key mappings received a description by this. The actual
benefit is questionable. After all, it is at least some kind of
"documentation". Especially for rarely used functionality.