chore: translate all key mappings to Lua
This is a wrap-up task to clean up the mix of VimScript and Lua key map
definitions. It's not like Lua is always the best. But having it aligned
is a good idea.

Some mappings have already been removed or slightly improved on the go.
A future task will be to iterate over all mappings again to clean them
up by actual usage pattern.

This also continues to add a description property for documentation
purposes on every key mapping.

The usage of `vim.api.nvim_set_keymaps` have been resolved in advantage
of `vim.keymap.set`, which is recommended.