chore: remove usage of config and setup for plugins
This is another commit in the series to use less external plugin manager
features. The affected plugins were almost all start plugins without
lazy loading. So a proper usage of start packages solved the "issue" for
many.
Setup functions/commands are a little more complicated, as they need to
be loaded before the actual package. This has been partially solved,
partially resolved. In general, this is a hint that the `after`
directory usage needs to be re-thought.