chore: introduce and apply custom lazy loading by events
Another commit in the series to continuously remove the dependency of an
external plugin manager. This time, the lazy loading by events has been
replaced. The logic is super trivial.

This puts some emphasize on an interesting development on how the start
and opt directories for plugins evolve. Some future improvements are in
the making.