chore: resolve plugin requires usage
This is another commit in the series to remove dependencies to an
external plugin manager. That includes the `requires` feature. It
heavily simplifies this, by just adding multiple plugins after the
"main" plugin in a plugin registration script. The loading was already
solved with a manual approach as well. So this fits very nicely.