chore(plugins): move and rename scripts to install plugins
First of all, the scripts have been renamed to `plugin_specification`.
I consider this a better naming than `register_plugin`. The term of
a "specification" comes from the latest development of `vim.pack`. As it
is currently the anticipated goal to migrate to this solution in future,
this makes even more sense.

Moreover, these scripts have been moved into the start directory of the
package. It is still a script in the root of each plugin, so it has no
effect until manually sourced. That is the same as before.
However, with this change there comes more structure towards start and
opt plugins. Every plugin is start per default unless establishing some
lazy loading. In best case, plugins are written in a way so that custom
lazy loading is not necessary. But if there is, an opt plugin will
complement it. This is currently always the case for lazy loaded plugins
that require to call a `setup` function.