feat(plugin/lsp): extract also inlay hints with Lua code
This was the only formerly packaged lsp plugin, which did not contain
a plugin folder only, but also some complementing Lua directory.
The plugin scripts have been moved to the top level `plugin` structure.
The Lua module has been moved to a new top level `lua` structure now.
This splits the related parts apart. Which is not 100% optimal, but it
does not justifies the overly complex structure.