feat(lsp): start to use brand new NeoVim LSP configuration
This extracts the Lua language server into the new top level `lsp`
folder structure of NeoVim. Furthermore, it adds a plugin script to
automatically enable all LSP clients, that are installed.
This approach was chosen over a `vim.lsp.start` approach, to take
advantage of the new `lsp` folder for a more obvious and clear
structure. The automatic enabling with an additional check ensures that
plenty of clients can be configured for various kind of projects and
filetypes. But it does not require all of them to be installed system
wide. It rather checks if the current "project" where NeoVim runs in has
a client available. I don't have the Lua language server installed
globally. But for example in the development environment for my NeoVim
configuration.