chore(treesitter): update to "main" branch
The leading NeoVim repository for treesitter queries switched to the
`main` branch. Though, it is not yet the default branch for the remote
repository.
This comes also with some breaking changes. Especially for the
integration of other treesitter based plugins. All relevant plugins got
updated. However, the nvim-treesitter-refactor plugin is not yet
compatible and got disabled. I consider to replace it with a custom
implementation, as I basically use a single feature only.