fix: (local)leader mappings
After the definition of the (local)leader variables was moved with the
options to a plugin script, some key mappings in other plugin scripts
stopped working. Turned out the loading order of the plugin scripts
affect this. Thereby, some key mappings using the leader worked, others
not. To resolve that issue, both variables are defined in the `init.lua`
again.