feat: adapt function related key mappings
For some reason, in the past I considered 'f' to be taken by some NeoVim
default text-object. But that is actually no the case. Therefore
I changed it, as it is more intuitive.
Notice that for example "parameters" are still called "arguments", as
'p' is actually a (partial) text-object by default for paragraphs.