feat(plugins/iswap.nvim): change key mapping
Avoid using the leader key here. It was basically "blocking" the usage
of the 'c' key mapping group from a semantic perspective. As the 'c'
stand for "change", I decided to align it with native change operator
mappings. The mapping for 'cs' is used by the vim-surround plugin. It is
a tight fit, but alright.